Fixes #21326: Move search cache job to jobs.py and onto the backend
Relocate SearchCacheJob into netbox/search/jobs.py, matching the per-app
job convention (core/jobs.py, extras/jobs.py); tasks.py was the lone
violation. The job is now thin: it delegates to the backend.

Move the deferred-update logic (the remove loop, the re-fetch +
atomic remove-then-cache loop, and the missing-schema skip) from the
free function update_search_cache onto CachedValueSearchBackend as
apply_deferred_updates(), with _is_missing_schema and the Postgres
SQLSTATE set as private members of that class. The CachedValue specifics
(_remove_by_id, cache(..., using=...)) are now reached only from within
the backend that owns them, rather than from the job and the inline
fallback reaching into backend internals. SearchBackend gains an abstract
apply_deferred_updates() (raising NotImplementedError like its siblings)
so the deferred contract is explicit for custom SEARCH_BACKEND classes.

deferred._flush now dispatches via SearchCacheJob (worker path) or
search_backend.apply_deferred_updates() (inline fallback). This renames
and relocates the backends/deferred/jobs import cycle but does not break
it; the single function-local import in _flush remains, and the clean fix
(extracting the search_backend singleton) is still tracked in #22485.
